Remove _get_client_option
Instead of using a custom reader method around config options such that
it falls back to a common set of defaults set the default value of those
client specific options to the default value from the group when the
options are registered.

This is required to allow keystoneclient to load session objects
directly from that group's config options.